socialmedia/socialmedia/requirements.txt
Django==4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.3.1
pytz